+++ platform-sdk/base-concurrent/src/main/java/module-info.java
@@ -3,6 +3,7 @@ module org.hiero.base.concurrent {
     exports org.hiero.base.concurrent;
     exports org.hiero.base.concurrent.atomic;
     exports org.hiero.base.concurrent.futures;
+    exports org.hiero.base.concurrent.interrupt;
     exports org.hiero.base.concurrent.locks;
     exports org.hiero.base.concurrent.locks.locked;
