+++ platform-sdk/base-concurrent/src/main/java/org/hiero/base/concurrent/AbstractTask.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.concurrent;
+package org.hiero.base.concurrent;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.concurrent.ForkJoinPool;
