+++ platform-sdk/base-concurrent/src/main/java/org/hiero/base/concurrent/internal/DefaultExecutorFactory.java
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.concurrent.internal;
+package org.hiero.base.concurrent.internal;
 
-import com.swirlds.common.concurrent.ExecutorFactory;
+import org.hiero.base.concurrent.ExecutorFactory;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.lang.Thread.UncaughtExceptionHandler;
