+++ platform-sdk/base-concurrent/src/main/java/org/hiero/base/concurrent/internal/DefaultForkJoinWorkerThreadFactory.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.concurrent.internal;
+package org.hiero.base.concurrent.internal;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
