+++ platform-sdk/base-concurrent/src/timingSensitive/java/org/hiero/base/concurrent/ExecutorFactoryTest.java
@@ -1,9 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.test.concurrent;
+package org.hiero.base.concurrent;
 
 import com.swirlds.base.context.Context;
-import com.swirlds.common.concurrent.ExecutorFactory;
-import com.swirlds.common.concurrent.internal.DefaultExecutorFactory;
+import org.hiero.base.concurrent.internal.DefaultExecutorFactory;
 import com.swirlds.logging.api.Logger;
 import com.swirlds.logging.api.Loggers;
 import com.swirlds.logging.test.fixtures.LoggingMirror;
