+++ platform-sdk/base-concurrent/src/timingSensitive/java/org/hiero/base/concurrent/internal/DefaultForkJoinWorkerThreadFactoryTest.java
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.test.concurrent;
+package org.hiero.base.concurrent.internal;
 
-import com.swirlds.common.concurrent.internal.DefaultForkJoinWorkerThreadFactory;
 import java.lang.Thread.UncaughtExceptionHandler;
 import java.util.concurrent.ForkJoinPool;
 import java.util.concurrent.ForkJoinPool.ForkJoinWorkerThreadFactory;
