+++ platform-sdk/base-concurrent/src/timingSensitive/java/org/hiero/base/concurrent/internal/ThreadLocalHandlingTest.java
@@ -1,8 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.test.concurrent;
+package org.hiero.base.concurrent.internal;
 
-import com.swirlds.common.concurrent.ExecutorFactory;
-import com.swirlds.common.concurrent.internal.DefaultExecutorFactory;
+import org.hiero.base.concurrent.ExecutorFactory;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.lang.Thread.UncaughtExceptionHandler;
 import java.util.ArrayList;
